authenticators/extra: remove issuer from email
Doing this means the header comes out like:

	aristaEmail: apd@arista.com:cv-infra

This is not useful to us at this point in the authz process. I can't
think of a reason an app on a cluster would need to know the source
of the JWT. Even if it did, using the pre-existing field is not a
good idea either.