fix: correctly delete cookie when revoking session
rfc6265 states: "to remove a cookie, the server returns a Set-Cookie header
with an expiration date in the past.  The server will be successful
in removing the cookie only if the Path and the Domain attribute in
the Set-Cookie header match the values used when the cookie was
created."

Before this PR, the domain is defined when the cookie is set,
but is not defined when the cookie is deleted. This PR
addresses this, and results in the cookie actually being
deleted