feat: Do not redirect robot-users to the login portal
Today when you make a request using expired tokens, oidc will
redirect you to a login portal. This is nice for humans who can
see the portal and decide to re-auth, but this makes things hard
for robot clients. The difficulty for robot clients takes two forms:
first, they are unable to use the login portal that we redirect
them two; second, many programs will interpret a 302 response to
mean that authentication was successful.

With this change, we will give these robot-clients a 401 response
which explicitly tells them that they are not authorized. Clients
should infer that they need to re-authenticate.

I've chosen the `X-Requested-With` header based on the TODO comment.
As far as I understand, the only common value for this header is
`XMLHttpRequest`, but I am just checking for a nonempty value here.